Added service counter to add id number in collections dynamically

+++ src/main/java/com/mcompany/inventory/service/DbRepository.java
@@ -7,7 +7,7 @@ interface DbRepository<T>  {
 
     void create(T object);
     void update(T object);
-    void delete(Long id);
-    T findOne(Long id);
+    void delete(int id);
+    T findOne(int id);
     List<T> findAll();
 }
